chapter-03-ai-agent-integration/langchain/requirements.txt
pyspiffe==0.7.0
langchain==0.1.0
langchain-openai==0.0.2
openai>=1.0.0
requests>=2.31.0
cryptography>=41.0.0
